KpyTou_BACKEND

+++ glasses_api/views/AuthViews.py
@@ -1,23 +1,20 @@
-from django.contrib.auth import authenticate, login, logout
-from django.contrib.auth.hashers import make_password
+from django.contrib.auth import authenticate, logout
 from django.http import HttpResponse
-from django.views.decorators.csrf import csrf_exempt
-from rest_framework.viewsets import ModelViewSet
 from rest_framework.response import Response
 from rest_framework import status
+from rest_framework.viewsets import ModelViewSet
 from rest_framework.permissions import AllowAny
-from rest_framework.decorators import permission_classes, authentication_classes, api_view
+from rest_framework.decorators import permission_classes
 from drf_yasg.utils import swagger_auto_schema
 
-import json
+from datetime import timedelta
+
 import uuid
-import redis
-from BACKEND.settings import REDIS_HOST, REDIS_PORT
 
-from ..models import User
-from ..serializers import UserSerializer
-from ..permissions import *
-
+from ..serializers import *
+from rest_framework.decorators import api_view
+from glasses_api.permissions import *
+from ..services import get_session
 
 session_storage = redis.StrictRedis(host=REDIS_HOST, port=REDIS_PORT)
 
@@ -57,59 +54,56 @@ class UserViewSet(ModelViewSet):
         return Response({'status': 'Error', 'error': serializer.errors}, status=status.HTTP_400_BAD_REQUEST)
 
 
-@permission_classes([AllowAny])
-@authentication_classes([])
-@csrf_exempt
+@api_view(['Post'])
+@permission_classes([IsAuthenticated])
+def check(request):
+    session_id = request.headers.get("authorization")
+
+    if session_storage.get(session_id):
+        user = User.objects.get(username=session_storage.get(session_id).decode('utf-8'))
+        serializer = UserSerializer(user, many=False)
+        return Response(serializer.data, status=status.HTTP_200_OK)
+
+    return Response(status=status.HTTP_403_FORBIDDEN)
+
+
 @swagger_auto_schema(method='post', request_body=UserSerializer)
 @api_view(['Post'])
+@permission_classes([AllowAny])
 def login_view(request):
     username = request.data["username"]
     password = request.data["password"]
-    user = authenticate(request, username=username, password=password)
+    user: User = authenticate(request, username=username, password=password)
+
     if user is not None:
         random_key = str(uuid.uuid4())
         session_storage.set(random_key, username)
 
-        response = HttpResponse(json.dumps({
-            'status': 'success',
-            'pk': user.pk,
-            'username': user.username,
-            'is_moderator': user.is_moderator,
-            'active_order': user.active_order,
-            'is_authenticated': user.is_authenticated,
-        }))
-        response.set_cookie("session_id", random_key)
+        data = {
+            "session_id": random_key,
+            "user_id": user.pk,
+            "username": user.username,
+            "is_moderator": user.is_moderator
+        }
 
+        response = Response(data, status=status.HTTP_201_CREATED)
+        response.set_cookie("session_id", random_key, httponly=False, expires=timedelta(days=1))
         return response
     else:
-        return HttpResponse("{'status': 'error', 'error': 'login failed'}")
+        return HttpResponse(status=status.HTTP_403_FORBIDDEN)
 
 
-@permission_classes([AllowAny])
-@authentication_classes([])
-@csrf_exempt
 @swagger_auto_schema(method='post')
 @api_view(['Post'])
+@permission_classes([AllowAny])
 def logout_view(request):
-    try:
-            ssid = request.COOKIES["session_id"]
-    except:
-        return HttpResponse("{'status': 'error', 'error': 'logout failed'}")
-        
+    ssid = get_session(request)
+    if ssid is None:
+        return Response(status=status.HTTP_401_UNAUTHORIZED)
+
     session_storage.delete(ssid)
 
     logout(request._request)
-    response = HttpResponse("{'status': 'success'}")
+    response = HttpResponse(status=status.HTTP_200_OK)
     response.delete_cookie("session_id")
     return response
-
-
-@api_view(['Post'])
-@permission_classes([AllowAny])
-def check(request):
-    session_id = request.headers.get("authorization")
-    if (session_storage.get(session_id)):
-        user = User.objects.get(username=session_storage.get(session_id).decode('utf-8'))
-        serializer = UserSerializer(user)
-        return Response(serializer.data, status=status.HTTP_200_OK)
-    return Response(status=status.HTTP_403_FORBIDDEN)
